Perform rounding on floating points before casting to uint64_t.

+++ cantranslator/canwrite.cpp
@@ -45,7 +45,7 @@ uint64_t encodeCanSignal(CanSignal* signal, float value) {
 }
 
 uint64_t encodeCanSignal(CanSignal* signal, float value, uint64_t data) {
-    uint64_t rawValue = (value - signal->offset) / signal->factor;
+    float rawValue = (value - signal->offset) / signal->factor;
     if(rawValue > 0) {
         // round up to avoid losing precision when we cast to an int
         rawValue += 0.5;
